requirement.txt
alembic==1.6.5
aniso8601==9.0.1
captcha==0.3
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==3.0.1
Flask-Pydantic==0.9.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
Pillow==9.4.0
psutil==5.8.0
pydantic==1.9.1
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.18
typing_extensions==4.2.0
Werkzeug==2.0.1
